+++ CHANGELOG.md
@@ -14,6 +14,18 @@ This changelog also contains important changes in dependencies.
 - `usvg_tree::Mask::kind`
 - (rosvgtree) New SVG 2 mask attributes.
 
+### Changed
+- `BackgroundImage` and `BackgroundAlpha` filter inputs will produce the same output
+  as `SourceGraphic` and `SourceAlpha` respectively.
+
+### Removed
+- `enable-background` support. This feature was never supported by browsers
+  and was deprecated in SVG 2. To my knowledge, only Batik has a good support of it.
+  Also, it's performance nightmare, which caused multiple issues in resvg already.
+- `usvg_tree::EnableBackground`
+- `usvg_tree::Group::enable_background`
+- `usvg_tree::NodeExt::filter_background_start_node`
+
 ### Fixed
 - Improve rectangular clipping anti-aliasing quality.
 - Mask's RGB to Luminosity converter was ignoring premultiplied alpha.